refactor(interfaces): reorganize domain interfaces and update definitions
