+++ src/mail/Mail.ts
@@ -1,7 +1,5 @@
 export interface Mail {
   address(): Promise<string>
-
   password(): Promise<string>
-
   delete(): Promise<void>
 }
